+++ src/main/java/com/groom/sumbisori/domain/file/dto/request/FileInfo.java
@@ -1,4 +1,4 @@
-package com.groom.sumbisori.domain.file.dto;
+package com.groom.sumbisori.domain.file.dto.request;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
