+++ src/main/java/com/groom/sumbisori/domain/file/dto/request/PreSignedUrlRequest.java
@@ -1,4 +1,4 @@
-package com.groom.sumbisori.domain.file.dto;
+package com.groom.sumbisori.domain.file.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
